healthtrack/requirements.txt
streamlit==1.29.0
pandas>=1.3.0
python-dateutil>=2.8.2
SQLAlchemy>=1.4.0
Pillow>=9.0.0